Fixes all ports to be `uint16`
Some code was incorrectly using `int16` for ports (which truncates) it or `int` (which is ok but unncessary). This PR addresses most of it. However, the ontology `ports` attribute in a network service also incorrectly uses `int16` and this needs to be fixed by `owl2java`

+++ rest/rest.go
@@ -93,7 +93,7 @@ var (
 	DefaultAllowedMethods = []string{"GET", "POST", "PUT", "DELETE"}
 
 	// DefaultAPIHTTPPort specifies the default port for the REST API.
-	DefaultAPIHTTPPort int16 = 8080
+	DefaultAPIHTTPPort uint16 = 8080
 )
 
 func init() {
@@ -168,7 +168,7 @@ func WithEmbeddedOAuth2Server(keyPath string, keyPassword string, saveOnCreate b
 
 // RunServer starts our REST API. The REST API is a reverse proxy using grpc-gateway that
 // exposes certain gRPC calls as RESTful HTTP methods.
-func RunServer(ctx context.Context, grpcPort int, httpPort int, serverOpts ...ServerConfigOption) (err error) {
+func RunServer(ctx context.Context, grpcPort uint16, httpPort uint16, serverOpts ...ServerConfigOption) (err error) {
 	ctx, cancel := context.WithCancel(ctx)
 	defer cancel()
